Actualizacion codigos

+++ Figura.java
@@ -2,5 +2,5 @@ public abstract class Figura {
     
     public abstract double calcularArea();
 
-    public abstract int numeroLados();
+    public abstract double numeroLados();
 }
